+++ test/Stormpath.Configuration.Test/DefaultConfigTestCases/JsonTestCase.cs
@@ -53,16 +53,10 @@ namespace Stormpath.Configuration.Test.DefaultConfigTestCases
       ""password"": null
     }
   },
+  ""apiToken"": ""okta_apiToken"",
+  ""org"": ""okta_org"",
   ""application"": {
-    ""name"": null,
-    ""href"": null
-  },
-  ""okta"": {
-    ""apiToken"": ""okta_apiToken"",
-    ""org"": ""okta_org"",
-    ""application"": {
-      ""id"": ""okta_application_id""
-    }
+    ""id"": ""okta_application_id""
   },
   ""web"": {
     ""serverUri"": null,