+++ test/Stormpath.Configuration.Test/DefaultConfigTestCases/YamlTestCase.cs
@@ -49,14 +49,10 @@ client:
     host: null
     username: null
     password: null
+apiToken: okta_apiToken
+org: okta_org
 application:
-  name: null
-  href: null
-okta:
-  apiToken: okta_apiToken
-  org: okta_org
-  application:
-    id: okta_application_id
+  id: okta_application_id
 web:
   serverUri: null
   basePath: ""/""