+++ test/Stormpath.Configuration.Test/ModifiedConfigTestCases/JsonTestCase.cs
@@ -59,16 +59,10 @@ namespace Stormpath.Configuration.Test.ModifiedConfigTestCases
       ""password"": ""bar""
     }
   },
+  ""apiToken"": ""foobarApiToken"",
+  ""org"": ""https://dev-12345.oktapreview.com"",
   ""application"": {
-    ""name"": ""Lightsabers Galore"",
-    ""href"": ""https://api.foo.com/v1/applications/foo""
-  },
-  ""okta"": {
-    ""apiToken"": ""foobarApiToken"",
-    ""org"": ""https://dev-12345.oktapreview.com"",
-    ""application"": {
-      ""id"": ""LightsabersGalore.app.foo""
-    }
+    ""id"": ""LightsabersGalore.app.foo""
   },
   ""web"": {
     ""serverUri"": ""https://localhost:9000"",